feat(admin): replace inline copy link with PublicLinkChip pill
Implements the "Public link" pill from the Forms IA design ("Guided
empty state" V1 Tabs section): a rounded chip with label divider,
truncated monospace URL, copy button (success-tinted on click), and an
external-link affordance. Removes the older CopyLinkField, which had
only one consumer.
